Change type bundle item score and penalty from Integer to Double to follow Sandalphon (with DB migration) (#122)

+++ judgels-backends/sandalphon/sandalphon-api/src/main/java/judgels/sandalphon/api/problem/bundle/MultipleChoiceItemConfig.java
@@ -8,8 +8,8 @@ import org.immutables.value.Value;
 @Value.Immutable
 @JsonDeserialize(as = ImmutableMultipleChoiceItemConfig.class)
 public interface MultipleChoiceItemConfig extends ItemConfig {
-    int getScore();
-    int getPenalty();
+    double getScore();
+    double getPenalty();
     List<Choice> getChoices();
 
     @Value.Immutable
